--- a/span.go
+++ b/span.go
@@ -9,17 +9,17 @@ import (
 	"go.elastic.co/apm/stacktrace"
 )
 
-// droppedSpanPool holds *Spans which are used when the span
-// is created for a nil or non-sampled Transaction, or one
-// whose max spans limit has been reached.
-var droppedSpanPool sync.Pool
+// droppedSpanDataPool holds *SpanData which are used when the span
+// is created for a nil or non-sampled Transaction, or one whose max
+// spans limit has been reached.
+var droppedSpanDataPool sync.Pool
 
 // StartSpan starts and returns a new Span within the transaction,
 // with the specified name, type, and optional parent span, and
 // with the start time set to the current time.
 //
-// StartSpan always returns a non-nil Span. Its End method must
-// be called when the span completes.
+// StartSpan always returns a non-nil Span, with a non-nil SpanData
+// field. Its End method must be called when the span completes.
 //
 // StartSpan is equivalent to calling StartSpanOptions with
 // SpanOptions.Parent set to the trace context of parent if
@@ -27,7 +27,13 @@ var droppedSpanPool sync.Pool
 func (tx *Transaction) StartSpan(name, spanType string, parent *Span) *Span {
 	var parentTraceContext TraceContext
 	if parent != nil {
+		parent.mu.RLock()
+		if parent.ended() {
+			parent.mu.RUnlock()
+			return newDroppedSpan()
+		}
 		parentTraceContext = parent.TraceContext()
+		parent.mu.RUnlock()
 	}
 	return tx.StartSpanOptions(name, spanType, SpanOptions{
 		Parent: parentTraceContext,
@@ -40,13 +46,24 @@ func (tx *Transaction) StartSpan(name, spanType string, parent *Span) *Span {
 // StartSpan always returns a non-nil Span. Its End method must
 // be called when the span completes.
 func (tx *Transaction) StartSpanOptions(name, spanType string, opts SpanOptions) *Span {
-	if tx == nil || !tx.Sampled() {
+	if tx == nil {
 		return newDroppedSpan()
 	}
-	tx.mu.Lock()
+
+	// Prevent tx from being ended while we're starting a span.
+	tx.mu.RLock()
+	defer tx.mu.RUnlock()
+
+	if tx.ended() || !tx.traceContext.Options.Recorded() {
+		return newDroppedSpan()
+	}
+
+	// Guard access to spansCreated, spansDropped, and rand.
+	tx.TransactionData.mu.Lock()
+	defer tx.TransactionData.mu.Unlock()
+
 	if tx.maxSpans > 0 && tx.spansCreated >= tx.maxSpans {
 		tx.spansDropped++
-		tx.mu.Unlock()
 		return newDroppedSpan()
 	}
 	transactionID := tx.traceContext.Span
@@ -65,7 +82,6 @@ func (tx *Transaction) StartSpanOptions(name, spanType string, opts SpanOptions)
 	binary.LittleEndian.PutUint64(span.traceContext.Span[:], tx.rand.Uint64())
 	span.stackFramesMinDuration = tx.spanFramesMinDuration
 	tx.spansCreated++
-	tx.mu.Unlock()
 	return span
 }
 
@@ -100,115 +116,6 @@ func (t *Tracer) StartSpan(name, spanType string, transactionID SpanID, opts Spa
 	return span
 }
 
-func (t *Tracer) startSpan(name, spanType string, transactionID SpanID, opts SpanOptions) *Span {
-	span, _ := t.spanPool.Get().(*Span)
-	if span == nil {
-		span = &Span{
-			tracer:   t,
-			Duration: -1,
-		}
-	}
-	span.Name = name
-	span.Type = spanType
-	span.traceContext = opts.Parent
-	span.parentID = opts.Parent.Span
-	span.transactionID = transactionID
-	span.timestamp = opts.Start
-	return span
-}
-
-// Span describes an operation within a transaction.
-type Span struct {
-	tracer                 *Tracer // nil if span is dropped
-	traceContext           TraceContext
-	parentID               SpanID
-	transactionID          SpanID
-	stackFramesMinDuration time.Duration
-	timestamp              time.Time
-
-	Name     string
-	Type     string
-	Duration time.Duration
-	Context  SpanContext
-
-	stacktrace []stacktrace.Frame
-}
-
-func newDroppedSpan() *Span {
-	span, _ := droppedSpanPool.Get().(*Span)
-	if span == nil {
-		span = &Span{}
-	}
-	return span
-}
-
-func (s *Span) reset() {
-	*s = Span{
-		tracer:     s.tracer,
-		Context:    s.Context,
-		Duration:   -1,
-		stacktrace: s.stacktrace[:0],
-	}
-	s.Context.reset()
-	s.tracer.spanPool.Put(s)
-}
-
-// TraceContext returns the span's TraceContext.
-func (s *Span) TraceContext() TraceContext {
-	return s.traceContext
-}
-
-// SetStacktrace sets the stacktrace for the span,
-// skipping the first skip number of frames,
-// excluding the SetStacktrace function.
-func (s *Span) SetStacktrace(skip int) {
-	if s.Dropped() {
-		return
-	}
-	s.stacktrace = stacktrace.AppendStacktrace(s.stacktrace[:0], skip+1, -1)
-}
-
-// Dropped indicates whether or not the span is dropped, meaning it will not
-// be included in any transaction. Spans are dropped by Transaction.StartSpan
-// if the transaction is nil, non-sampled, or the transaction's max spans
-// limit has been reached.
-//
-// Dropped may be used to avoid any expensive computation required to set
-// the span's context.
-func (s *Span) Dropped() bool {
-	return s.tracer == nil
-}
-
-// End marks the s as being complete; s must not be used after this.
-//
-// If s.Duration has not been set, End will set it to the elapsed time
-// since the span's start time.
-func (s *Span) End() {
-	if s.Dropped() {
-		droppedSpanPool.Put(s)
-		return
-	}
-	if s.Duration < 0 {
-		s.Duration = time.Since(s.timestamp)
-	}
-	if len(s.stacktrace) == 0 && s.Duration >= s.stackFramesMinDuration {
-		s.SetStacktrace(1)
-	}
-	s.enqueue()
-}
-
-func (s *Span) enqueue() {
-	select {
-	case s.tracer.spans <- s:
-	default:
-		// Enqueuing a span should never block.
-		s.tracer.statsMu.Lock()
-		s.tracer.stats.SpansDropped++
-		s.tracer.statsMu.Unlock()
-		s.reset()
-	}
-}
-
 // SpanOptions holds options for Transaction.StartSpanOptions and Tracer.StartSpan.
 type SpanOptions struct {
 	// Parent, if non-zero, holds the trace context of the parent span.
@@ -227,4 +134,169 @@ type SpanOptions struct {
 	// transaction timestamp. Calculating the timstamp in this way will ensure
 	// monotonicity of events within a transaction.
 	Start time.Time
+}
+
+func (t *Tracer) startSpan(name, spanType string, transactionID SpanID, opts SpanOptions) *Span {
+	sd, _ := t.spanDataPool.Get().(*SpanData)
+	if sd == nil {
+		sd = &SpanData{
+			tracer:   t,
+			Duration: -1,
+		}
+	}
+	span := &Span{SpanData: sd}
+	span.Name = name
+	span.Type = spanType
+	span.traceContext = opts.Parent
+	span.parentID = opts.Parent.Span
+	span.transactionID = transactionID
+	span.timestamp = opts.Start
+	return span
+}
+
+func newDroppedSpan() *Span {
+	span, _ := droppedSpanDataPool.Get().(*Span)
+	if span == nil {
+		span = &Span{SpanData: &SpanData{}}
+	}
+	return span
+}
+
+// Span describes an operation within a transaction.
+type Span struct {
+	mu sync.RWMutex
+
+	// SpanData holds the span data. This field is set to nil when
+	// the span's End method is called.
+	*SpanData
+}
+
+// TraceContext returns the span's TraceContext.
+func (s *Span) TraceContext() TraceContext {
+	if s == nil {
+		return TraceContext{}
+	}
+	s.mu.RLock()
+	defer s.mu.RUnlock()
+	if s.ended() {
+		return TraceContext{}
+	}
+	return s.traceContext
+}
+
+// SetStacktrace sets the stacktrace for the span,
+// skipping the first skip number of frames,
+// excluding the SetStacktrace function.
+func (s *Span) SetStacktrace(skip int) {
+	if s == nil {
+		return
+	}
+	s.mu.RLock()
+	defer s.mu.RUnlock()
+	if s.ended() || s.dropped() {
+		return
+	}
+	s.SpanData.setStacktrace(skip + 1)
+}
+
+// Dropped indicates whether or not the span is dropped, meaning it will not
+// be included in any transaction. Spans are dropped by Transaction.StartSpan
+// if the transaction is nil, non-sampled, or the transaction's max spans
+// limit has been reached.
+//
+// Dropped may be used to avoid any expensive computation required to set
+// the span's context.
+func (s *Span) Dropped() bool {
+	if s == nil {
+		return true
+	}
+	s.mu.RLock()
+	dropped := s.ended() || s.dropped()
+	s.mu.RUnlock()
+	return dropped
+}
+
+// End marks the s as being complete; s must not be used after this.
+//
+// If s.Duration has not been set, End will set it to the elapsed time
+// since the span's start time.
+func (s *Span) End() {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	if s.ended() || s.dropped() {
+		droppedSpanDataPool.Put(s.SpanData)
+		return
+	}
+	if s.Duration < 0 {
+		s.Duration = time.Since(s.timestamp)
+	}
+	if len(s.stacktrace) == 0 && s.Duration >= s.stackFramesMinDuration {
+		s.setStacktrace(1)
+	}
+	s.SpanData.enqueue()
+	s.SpanData = nil
+}
+
+func (s *Span) ended() bool {
+	return s.SpanData == nil
+}
+
+// SpanData holds the details for a span, and is embedded inside Span.
+// When a span is ended or discarded, its SpanData field will be set
+// to nil.
+type SpanData struct {
+	tracer                 *Tracer // nil if span is dropped
+	traceContext           TraceContext
+	parentID               SpanID
+	transactionID          SpanID
+	stackFramesMinDuration time.Duration
+	timestamp              time.Time
+
+	// Name holds the span name, initialized with the value passed to StartSpan.
+	Name string
+
+	// Type holds the span type, initialized with the value passed to StartSpan.
+	Type string
+
+	// Duration holds the span duration, initialized to -1.
+	//
+	// If you do not update Duration, calling Span.End will calculate the
+	// duration based on the elapsed time since the span's start time.
+	Duration time.Duration
+
+	// Context describes the context in which span occurs.
+	Context SpanContext
+
+	stacktrace []stacktrace.Frame
+}
+
+func (s *SpanData) setStacktrace(skip int) {
+	s.stacktrace = stacktrace.AppendStacktrace(s.stacktrace[:0], skip+1, -1)
+}
+
+func (s *SpanData) dropped() bool {
+	return s.tracer == nil
+}
+
+func (s *SpanData) enqueue() {
+	select {
+	case s.tracer.spans <- s:
+	default:
+		// Enqueuing a span should never block.
+		s.tracer.statsMu.Lock()
+		s.tracer.stats.SpansDropped++
+		s.tracer.statsMu.Unlock()
+		s.reset()
+	}
+}
+
+func (s *SpanData) reset() {
+	*s = SpanData{
+		tracer:     s.tracer,
+		Context:    s.Context,
+		Duration:   -1,
+		stacktrace: s.stacktrace[:0],
+	}
+	s.Context.reset()
+	s.tracer.spanDataPool.Put(s)
 }