Transaction, Span, and Error are no longer pooled.
Instead, new *Data types are introduced which are
pooled, and those are embedded within their related
top-level object types. We add a read-write mutex
to the Transaction and Span types to guard access
to those such that a concurrent StartSpan and
Transaction.End will not panic.

This change introduces an additional allocation for
each of StartTransaction, StartSpan, and NewError.
The time overhead (measured in nanoseconds) is not
enough to warrant living with the lack of safety.

It is now safe to call End() multiple times, and
to call StartSpan with an ended transaction or parent
span. When StartSpan is called with an ended
transaction, a dropped span will be created.

Closes #317